Add plugin name
