+++ BsddRevitPlugin.Common/PackageContents.xml
@@ -1,0 +1,21 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<ApplicationPackage SchemaVersion="1.0"
+	ProductType="Application"
+	Name="BsddRevitPlugin"
+	AppVersion="0.2.0"
+    Description="Revit Template Addin"
+    SupportedLocales="Enu"
+    Author=""
+    AutodeskProduct="Revit"
+    FriendlyVersion="0.2.0"
+    AppNameSpace="appstore.exchange.autodesk.com">
+
+	<CompanyDetails Name="" />
+
+	<RuntimeRequirements OS="Win32|Win64" Platform="Revit" SeriesMin="R2020" SeriesMax="R2020" />
+
+	<Components Description="2023 parts">
+		<RuntimeRequirements OS="Win64" Platform="Revit" SeriesMin="R2023" SeriesMax="R2023" />
+		<ComponentEntry AppName="BsddRevitPlugin" Version="0.1.0" ModuleName="./Contents/2023/BsddRevitPlugin.addin" AppDescription="Add data from the buildingSMART Data Dictionary to your Revit models" />
+	</Components>
+</ApplicationPackage>
