+++ BsddRevitPlugin.Installer/Installer.iss
@@ -1,0 +1,56 @@
+; Script generated by the Inno Setup Script Wizard.
+; SEE THE DOCUMENTATION FOR DETAILS ON CREATING INNO SETUP SCRIPT FILES!
+
+#define MyAppName "bSDD Revit plugin"
+#define MyAppVersion "1.0.0"
+#define MyAppPublisher ""
+
+[Setup]
+; NOTE: The value of AppId uniquely identifies this application. Do not use the same AppId value in installers for other applications.
+; (To generate a new GUID, click Tools | Generate GUID inside the IDE.)
+AppId={{91763ED0-AB3D-47E3-9C9B-855A5645CA7E}
+AppName={#MyAppName}
+#define installerPath "C:\TEMP\installers\"  ; TODO: Decide where this goes later, maybe in Program Files (using e.g., {commonpf64})
+AppVersion={#MyAppVersion}
+AppVerName={#MyAppName} {#MyAppVersion}
+AppPublisher={#MyAppPublisher}
+DefaultDirName={#installerPath}
+DisableDirPage=yes
+DefaultGroupName=Revit Template Group
+DisableProgramGroupPage=yes
+
+#define Revit2023 "\Autodesk\ApplicationPlugins\BsddRevitPlugin.bundle\Contents\2023\"
+; We defined the path above similarly to e-verse's approach, which works just fine -- however we can also consider just
+; doing it the way we did it before, e.g. (note: we append the appdata dir path below with {userappdata}:
+; # define Revit2023 "\Autodesk\Revit\Addins\2023\"
+
+; TODO: Add SignTool setup once we have a license
+; Uncomment the following line to run in non administrative install mode (install for current user only.)
+PrivilegesRequired=lowest
+OutputDir=.\Output
+OutputBaseFilename=BsddRevitPluginAddin
+Compression=lzma
+SolidCompression=yes
+SetupIconFile="..\BsddRevitPlugin.Resources\Images\Icons\ASRR_A_icon.ico"
+OutputManifestFile=Setup-Manifest.txt
+UninstallDisplayName="bSDD Revit plugin Uninstall"
+UninstallDisplayIcon="..\BsddRevitPlugin.Resources\Images\Icons\ASRR_A_icon.ico"
+WizardSmallImageFile="..\BsddRevitPlugin.Resources\Images\Icons\ASRR_A_icon.bmp"
+WizardStyle=modern
+
+[Languages]
+Name: "english"; MessagesFile: "compiler:Default.isl";
+
+[Messages]
+SetupWindowTitle =  Setup {#SetupSetting("AppName")} Version: {#SetupSetting("AppVersion")}
+
+[Files]
+Source: "..\BsddRevitPlugin.2023\bin\Release\*"; DestDir: "{userappdata}{#Revit2023}\"; Flags: ignoreversion recursesubdirs createallsubdirs
+Source: "..\BsddRevitPlugin.Common\BsddRevitPlugin.addin"; DestDir: "{userappdata}{#Revit2023}\"; Flags: ignoreversion recursesubdirs createallsubdirs
+
+Source: "..\BsddRevitPlugin.Common\PackageContents.xml"; DestDir: "{userappdata}\Autodesk\ApplicationPlugins\BsddRevitPlugin.bundle\"; Flags: ignoreversion
+
+; NOTE: Don't use "Flags: ignoreversion" on any shared system files
+
+[Icons]
+Name: "{group}\{cm:UninstallProgram,{#MyAppName}}"; Filename: "{uninstallexe}"
