+++ BsddRevitPlugin.Resources/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="NLog" version="5.2.4" targetFramework="net48" />
+</packages>
