--- RevitTemplate.2023/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="NLog" version="5.2.4" targetFramework="net48" />
-  <package id="Revit_All_Main_Versions_API_x64" version="2024.0.0" targetFramework="net48" />
-</packages>
