--- RevitTemplate.Common/RevitTemplate.addin
@@ -1,14 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<RevitAddIns>
-	<AddIn Type="Application">
-		<Name>RevitTemplate</Name>
-		<FullClassName>RevitTemplate.Common.Startup</FullClassName>
-		<Text>Sample Revit Addin</Text>
-		<Description>A simple add-in to display how we create Revit add-ins at ASRR</Description>
-		<VisibilityMode>AlwaysVisible</VisibilityMode>
-		<Assembly>RevitTemplate.dll</Assembly>
-		<AddInId>601FFC79-C4C3-4A5B-A0FE-CE89CEC7EE04</AddInId>
-		<VendorId>ADSK</VendorId>
-		<VendorDescription>ASRR B.V., www.asrr.nl</VendorDescription>
-	</AddIn>
-</RevitAddIns>
