--- RevitTemplate.Installer/Installer.iss
@@ -1,56 +1,0 @@
-; Script generated by the Inno Setup Script Wizard.
-; SEE THE DOCUMENTATION FOR DETAILS ON CREATING INNO SETUP SCRIPT FILES!
-
-#define MyAppName "ASRR Revit Template Addin"
-#define MyAppVersion "1.0.0"
-#define MyAppPublisher "ASRR"
-
-[Setup]
-; NOTE: The value of AppId uniquely identifies this application. Do not use the same AppId value in installers for other applications.
-; (To generate a new GUID, click Tools | Generate GUID inside the IDE.)
-AppId={{91763ED0-AB3D-47E3-9C9B-855A5645CA7E}
-AppName={#MyAppName}
-#define installerPath "C:\asrr\installers\"  ; TODO: Decide where this goes later, maybe in Program Files (using e.g., {commonpf64})
-AppVersion={#MyAppVersion}
-AppVerName={#MyAppName} {#MyAppVersion}
-AppPublisher={#MyAppPublisher}
-DefaultDirName={#installerPath}
-DisableDirPage=yes
-DefaultGroupName=Revit Template Group
-DisableProgramGroupPage=yes
-
-#define Revit2023 "\Autodesk\ApplicationPlugins\RevitTemplate.bundle\Contents\2023\"
-; We defined the path above similarly to e-verse's approach, which works just fine -- however we can also consider just
-; doing it the way we did it before, e.g. (note: we append the appdata dir path below with {userappdata}:
-; # define Revit2023 "\Autodesk\Revit\Addins\2023\"
-
-; TODO: Add SignTool setup once we have a license
-; Uncomment the following line to run in non administrative install mode (install for current user only.)
-PrivilegesRequired=lowest
-OutputDir=.\Output
-OutputBaseFilename=RevitTemplateAddin
-Compression=lzma
-SolidCompression=yes
-SetupIconFile="..\RevitTemplate.Resources\Images\Icons\ASRR_A_icon.ico"
-OutputManifestFile=Setup-Manifest.txt
-UninstallDisplayName="ASRR Revit Template Addin Uninstall"
-UninstallDisplayIcon="..\RevitTemplate.Resources\Images\Icons\ASRR_A_icon.ico"
-WizardSmallImageFile="..\RevitTemplate.Resources\Images\Icons\ASRR_A_icon.bmp"
-WizardStyle=modern
-
-[Languages]
-Name: "english"; MessagesFile: "compiler:Default.isl";
-
-[Messages]
-SetupWindowTitle =  Setup {#SetupSetting("AppName")} Version: {#SetupSetting("AppVersion")}
-
-[Files]
-Source: "..\RevitTemplate.2023\bin\Release\*"; DestDir: "{userappdata}{#Revit2023}\"; Flags: ignoreversion recursesubdirs createallsubdirs
-Source: "..\RevitTemplate.Common\RevitTemplate.addin"; DestDir: "{userappdata}{#Revit2023}\"; Flags: ignoreversion recursesubdirs createallsubdirs
-
-Source: "..\RevitTemplate.Common\PackageContents.xml"; DestDir: "{userappdata}\Autodesk\ApplicationPlugins\RevitTemplate.bundle\"; Flags: ignoreversion
-
-; NOTE: Don't use "Flags: ignoreversion" on any shared system files
-
-[Icons]
-Name: "{group}\{cm:UninstallProgram,{#MyAppName}}"; Filename: "{uninstallexe}"
