--- RevitTemplate.Installer/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="NLog" version="5.2.4" targetFramework="net48" />
-  <package id="Tools.InnoSetup" version="6.2.2" targetFramework="net48" />
-</packages>
